+++ src/renderer/src/components/generated-audio-display.tsx
@@ -2,35 +2,13 @@ import { MediaGetSchema } from '@common/db-schemas/media.schema'
 import { Download, Expand, Play, Square, Trash2 } from 'lucide-react'
 import { useEffect, useMemo, useRef, useState } from 'react'
 import { Button } from './ui/button'
-import { cn } from '@renderer/lib/utils'
 import { generateTailwindGradient } from '@renderer/lib/colors'
 import { Spinner } from './ui/spinner'
-
-/**
- * Animated waveform bars — purely decorative / visual indicator.
- * When playing, bars animate; when paused they freeze.
- */
-const Waveform = ({ playing }: { playing: boolean }) => (
-  <div className="flex items-center justify-center gap-[3px] h-14 w-full">
-    {Array.from({ length: 28 }).map((_, i) => (
-      <span
-        key={i}
-        className={cn(
-          'inline-block w-[3px] rounded-full bg-foreground/30 origin-center',
-          playing && 'animate-waveform'
-        )}
-        style={{
-          height: `${20 + Math.sin(i * 0.8) * 14 + Math.cos(i * 1.3) * 10}px`,
-          animationDelay: playing ? `${(i * 60) % 400}ms` : '0ms'
-        }}
-      />
-    ))}
-  </div>
-)
+import Waveform from './waveform'
 
 /**
  * Card component to display a single generated audio item.
- * Shows waveform visualisation, prompt preview, voice label,
+ * Shows waveform visualization, prompt preview, voice label,
  * play/stop controls and delete / expand / download actions.
  */
 const GeneratedAudioDisplay = ({
